Rename nudity -> nsfw to more accurately reflect filter
